--- test1.py
@@ -1,17 +1,0 @@
-import test
-
-#test.add_customer("Jason", "Smith", "jason@smith.com")
-
-#test.delete_customer("17")
-
-customer_list = [
-    ("kdjfsgh","m,fhbg","nvkghdfg"),
-    ("ytvdsdf","ncjfgh","yr87gbf")
-    ]
-#test.add_customer_list(customer_list)
-
-#test.show_all()
-
-test.lookup_customer("first_name", "John")
-
-#test.print_column_names()
